# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Mahad (Absensi, Hafalan, dll)",
        "title": "Mahad Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/mentors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint untuk mengambil daftar mentor dengan pagination berdasarkan query parameter `page` dan `limit`",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mentor"
                ],
                "summary": "Mengambil daftar mentor dengan pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of mentors retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch mentors",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/mentors/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint untuk mengambil data mentor berdasarkan ID. Mahasantri yang dibimbing juga akan dimuat (Preload).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mentor"
                ],
                "summary": "Mengambil data mentor berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mentor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mentor data retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.MentorResponse"
                        }
                    },
                    "404": {
                        "description": "Mentor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch mentor",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint untuk memperbarui data mentor berdasarkan ID. Pastikan email mentor yang baru tidak digunakan oleh mentor lain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mentor"
                ],
                "summary": "Memperbarui data mentor berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mentor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data mentor yang akan diperbarui",
                        "name": "updateMentorRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMentorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mentor updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.MentorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or no changes detected",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mentor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "409": {
                        "description": "Email already in use",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to update mentor",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint untuk menghapus data mentor berdasarkan ID. Pastikan mentor yang ingin dihapus ada di database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mentor"
                ],
                "summary": "Menghapus mentor berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mentor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mentor deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mentor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to delete mentor",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.MahasantriResponse": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jurusan": {
                    "type": "string"
                },
                "mentor_id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                }
            }
        },
        "dto.MentorResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mahasantri": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MahasantriResponse"
                    }
                },
                "mahasantri_count": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateMentorRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}